+++ CHANGELOG.md
@@ -17,6 +17,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ---
 
+## [1.0.13] - 2026-02-20
+
+### Added
+- Parameters for browser vision
+
+### Changed
+- Enhance the browser vision features
+
+### Fixed
+- N/A
+
+---
+
 ## [1.0.12] - 2026-02-20
 
 ### Added